EDC2 Renamed Packages
