+++ book-info/src/main/java/com/edc/pps/info/model/Library.java
@@ -1,4 +1,4 @@
-package com.edc.pps.model;
+package com.edc.pps.info.model;
 
 import java.util.HashSet;
 import java.util.Set;
