+++ book-info/src/main/java/com/edc/pps/info/model/Rating.java
@@ -1,4 +1,4 @@
-package com.edc.pps.model;
+package com.edc.pps.info.model;
 
 public class Rating {
     private long ratingId;
